requirements.txt
pandas
openpyxl
xlrd
PyYAML
typer
jinja2
fastapi
uvicorn
python-multipart
beautifulsoup4
lxml
rapidfuzz
